Add an app-access log
This contains the controller class name and action that is triggered,
along with the time in millisecods that the request took to render
(excluding symfony bootup time). This allows us to monitor response
times on a per-route basis - for instance the findByPid route takes X on
average, while a category list takes Y on average. Thus we can identify
speed regressions within particular routes rather than across the whole
application.