Go back to using file-based logging for controller metrics
It turns out using cloudwatch takes a suprising amount of time to send
its metrics to the point of it impacting the amount of requests we can
handle overall. Thus we go back to using a log file for this info.

A later commit shall deal with configuring the logGroups and their
metrics logFilters for this file.